Extract module `kontrol.prove`
